New mixins for better SASS file structures
Using the BEM classification allows for a standardised approach to style class declarations. Having the new mixins of `element` and `modifier` will help cement this methodology.
Other proposed changes include a mixin for scrollbars, links and denoting the full criteria of a stylised element - instead of using the `:not()` pseudo-class.